+++ bugsnag-plugin-android-ndk/src/main/jni/utils/string.h
@@ -14,7 +14,7 @@ extern "C" {
 void bsg_strcpy(char *dst, const char *src) __asyncsafe;
 
 /**
- * Return the length of a string
+ * Return the length of a string, or 0 if the pointer is NULL.
  */
 size_t bsg_strlen(const char *str) __asyncsafe;